requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
beautifulsoup4==4.12.3
requests==2.32.3
tenacity==8.2.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
streamlit==1.38.0
sqlitedict==2.1.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.33
langchain-text-splitters==0.2.2
python-dotenv==1.0.1
numpy==1.26.4
pytest==8.3.2
google-generativeai==0.8.3
